Prefetch data on next.js export
We now download the data on static build and inject it into the site.

This means the page works without javascript, and has a faster initial
render time. We need to be careful about the size of the page as time
goes on, as well as making sure to update the build regularly.